requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.6
django-bootstrap-v5==1.0.11
django-jazzmin==2.6.0
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
Pillow==9.4.0
pytz==2022.7.1
requests==2.28.2
soupsieve==2.4.1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
